+++ src/main/java/com/example/store/repository/projection/OrderRow.java
@@ -1,4 +1,6 @@
-package com.example.store.entity;
+package com.example.store.repository.projection;
+
+import java.util.List;
 
 /**
  * This is used in the projection from the database of the orders to create the OrderDTO for the order request.
@@ -8,4 +10,5 @@ public interface OrderRow {
     String getDescription();
     Long getCustomerId();
     String getCustomerName();
+    List<ProductSummaryView> getProducts();
 }